orbit::component

Module props

Props module for component properties

Re-exports

pub use paste;

Modules

props ๐

Enhanced props system for Orbit UI components

Macros

define_props

Type-safe builder for props

define_props_advanced

Advanced props builder with validation and required field support

validate_field

Creates a validator that ensures a field meets a condition

Structs

CompositeValidator

A simple property validator that can be composed of multiple validators

PropsBuilder

Builder for constructing component props with validation

Enums

PropValidationError

Error indicating validation problems with props

PropValue

Represents a property that can be required or optional with default value

Traits

PropValidator

Trait for props validation