orbit::component::props

Module props

Enhanced props system for Orbit UI components

This module provides improved props functionality including:

Type-safe props definitions

Props validation

Default values

Required vs optional props

Re-exports

pub use paste;

Structs

CompositeValidator

A simple property validator that can be composed of multiple validators

PropsBuilder

Builder for constructing component props with validation

Enums

PropValidationError

Error indicating validation problems with props

PropValue

Represents a property that can be required or optional with default value

Traits

PropValidator

Trait for props validation