orbit::component

Trait Component

pub trait Component:
    Send
    + Sync
    + Any {
    type Props: Props + Clone;

    // Required methods
    fn create(props: Self::Props, context: Context) -> Self
       where Self: Sized;
    fn update(&mut self, props: Self::Props) -> Result<(), ComponentError>;
    fn render(&self) -> Result<Vec<Node>, ComponentError>;
    fn as_any(&self) -> &dyn Any;
    fn as_any_mut(&mut self) -> &mut dyn Any;

    // Provided methods
    fn initialize(&mut self) -> Result<(), ComponentError> { ... }
    fn mount(&mut self) -> Result<(), ComponentError> { ... }
    fn before_update(
        &mut self,
        _new_props: &Self::Props,
    ) -> Result<(), ComponentError> { ... }
    fn after_update(&mut self) -> Result<(), ComponentError> { ... }
    fn before_unmount(&mut self) -> Result<(), ComponentError> { ... }
    fn unmount(&mut self) -> Result<(), ComponentError> { ... }
}

Component trait - implemented by all UI components

Required Associated Types

type Props: Props + Clone

The props type for this component

Required Methods

fn create(props: Self::Props, context: Context) -> Self
where Self: Sized,

Create a new component instance

fn update(&mut self, props: Self::Props) -> Result<(), ComponentError>

Update component with new props

fn render(&self) -> Result<Vec<Node>, ComponentError>

Render component - returns child nodes

fn as_any(&self) -> &dyn Any

Convert to Any for downcasting

fn as_any_mut(&mut self) -> &mut dyn Any

Convert to mutable Any for downcasting

Provided Methods

fn initialize(&mut self) -> Result<(), ComponentError>

Initialize the component - called immediately after creation Use this for setting up initial state and registering lifecycle hooks

fn mount(&mut self) -> Result<(), ComponentError>

Mount component - called when component is first added to the tree

fn before_update( &mut self, _new_props: &Self::Props, ) -> Result<(), ComponentError>

Called before component updates with new props

fn after_update(&mut self) -> Result<(), ComponentError>

Called after the component has updated

fn before_unmount(&mut self) -> Result<(), ComponentError>

Called before component is unmounted

fn unmount(&mut self) -> Result<(), ComponentError>

Unmount component - called when component is removed from the tree