Trait orbit::OrbitBus

pub trait OrbitBus {
    // Required method
    fn init() -> Self;
}

An event bus that can be used to manage events in the Orbit system.

This is the central hub for all events in the system. It is used to subscribe and emit events. (see the disclaimer here)

Example

#[bus(Message, Player)]
struct Bus;
 
struct Player(String);

struct Message {
    id: u64,
    text: String,
    author: String,
}
 
struct MessageHandler;

impl OrbitEventHandler<Message> for MessageHandler {
    fn handle(&self, msg: &Message) {
        println!("'{}' by '{}' (#{})", msg.text, msg.author, msg.id);
    }
}

fn on_message(message: &Message) {
    println!("{}: {}", message.author, message.text);
}

let mut bus = Bus::init();
bus.sub(|player: &Player| {
    println!("Player joined: {}", player.0);
});
bus.sub(MessageHandler);
bus.sub(on_message);

bus.emit(Message {
    id: 1,
    text: "Hello, world!".to_string(),
    author: "Alice".to_string()
});
bus.emit(Player("Bob".to_string()));

Disclaimer

Actually, I lied. For subscribing events, you should use OrbitBusSubscribingManager, and for emitting events, you should use OrbitBusEmitingManager. If you want to be able to do both, use the OrbitBusManager trait. This trait is in reality only used to initialize the event bus. But the OrbitBusSubscribingManager and OrbitBusEmitingManager traits gets automatically implemented by the #[bus] macro, so you don’t have to worry about it. Unless of course, you want to implement it yourself raw. In that case, you should avoid the #[bus] macro at all costs, as it forbids you from adding anything to the Bus struct, which is required for manual implementation.

Required Methods

fn init() -> Self

Initializes the event bus.

Object Safety

This trait is not object safe.