Trait orbit::OrbitBusEmitingManager

pub trait OrbitBusEmitingManager<Event> {
    // Required method
    fn emit(&self, event: Event);
}

This trait is used to manage the emitting of events to the event bus.

This is a trait that must be implemented by the event bus, for it to properly emit events.

It will get automatically implemented if you use the #[bus] macro.

Required Methods

fn emit(&self, event: Event)

Emits an event to the event bus.

For examples see the OrbitBus trait.