Trait orbit::OrbitBusManager

pub trait OrbitBusManager<Event>: OrbitBusSubscribingManager<Event> + OrbitBusEmitingManager<Event> { }

This trait is used to both subscribe to and emit events to the event bus.

It will get automatically implemented if you use the #[bus] macro.

Example using the #[bus] macro

#[bus(Message, Player)]
struct Bus;

Example without the #[bus] macro

struct Bus {
    // Code to manage the event handlers goes here
}
struct Message(String);
impl OrbitBusSubscribingManager<Message> for Bus {
    fn sub<H>(&mut self, handler: H) where
        H: OrbitEventHandler<Message> + 'static
    {
        // Code to subscribe the handler to the event bus goes here
    }
}
impl OrbitBusEmitingManager<Message> for Bus {
    fn emit(&self, event: Message) {
        // Code to emit the event goes here
    }
}

Object Safety

This trait is not object safe.