Trait orbit::OrbitEventHandler

pub trait OrbitEventHandler<Event> {
    // Required method
    fn handle(&self, event: &Event);
}

This trait is used to handle events in the Orbit system.

This is a trait that all event handlers must implement. Note that this trait is automatically implemented for any function that takes a &Event and returns void.

Example

struct MessageEventHandler;

impl OrbitEventHandler<MessageEvent> for MessageEventHandler {
    fn handle(&self, event: &MessageEvent) {
        println!("Received message: {}", event.id);
    }
}

fn on_message(event: &MessageEvent) {
    println!("{}: {}", event.author, event.message);
}

let mut bus = Bus::init();
bus.on(MessageEventHandler);
bus.on(on_message);

Required Methods

fn handle(&self, event: &Event)

Function that gets invoked when the event is emitted.

Notice

This function should not block the thread, as it will block the entire event loop. It is recommended to spawn a new thread or use async/await to handle the event.