Crate orbit

Orbit Event System

Focused on simplicity and performance.

Example using the #[bus] macro

#[bus(Message, Player)]
struct Bus;
 
struct Player(String);

struct Message {
    id: u64,
    text: String,
    author: String,
}
 
struct MessageHandler;

impl OrbitEventHandler<Message> for MessageHandler {
    fn handle(&self, msg: &Message) {
        println!("'{}' by '{}' (#{})", msg.text, msg.author, msg.id);
    }
}

fn on_message(message: &Message) {
    println!("{}: {}", message.author, message.text);
}

let mut bus = Bus::init();
bus.sub(|player: &Player| {
    println!("Player joined: {}", player.0);
});
bus.sub(MessageHandler);
bus.sub(on_message);

bus.emit(Message {
    id: 1,
    text: "Hello, world!".to_string(),
    author: "Alice".to_string()
});
bus.emit(Player("Bob".to_string()));

For the architechture of the system, see the here

Traits

OrbitBus

An event bus that can be used to manage events in the Orbit system.

OrbitBusEmitingManager

This trait is used to manage the emitting of events to the event bus.

OrbitBusManager

This trait is used to both subscribe to and emit events to the event bus.

OrbitBusSubscribingManager

This trait is used to manage the subscribing of event handlers to the event bus.

OrbitEventHandler

This trait is used to handle events in the Orbit system.

Attribute Macros

bus

A macro to automatically generate the necessary code to manage an event bus.